requirements.txt
django
pillow
sorl-thumbnail
